Add core.js
A new JavaScript file, following the pattern used in www.ubuntu.com,
to host all core JavaScript for the site.

The idea here is that individual functions can be gradually migrated over from scratch.js as they're brought up-to-date.

Ultimately this file will initialise the "core" namespace, but for now that namespace is initialised by scratch.js, so I'm leaving the namespace code commented out.